# server/openapi.yaml
openapi: 3.1.0
info:
  title: Lumina Backend API
  version: 1.0.0
  description: |
    주차장 사고 증거 매칭 데모 MVP 백엔드 (BE_SPEC.md F0~F10).

    공통 규약
    - 모든 응답은 `{ data, meta }` 또는 `{ error: { code, message, retryable, details? }, meta: { requestId } }`.
    - 인증: `Authorization: Bearer <Supabase access_token>`. `/config`, `/health`, `/ready`, `/demo/*` 만 공개.
    - 시간은 ISO 8601 UTC 문자열. 금액은 KRW 정수.
    - 보험사 채택·정산은 Mock 이며 관련 DTO 는 항상 `mock: true` 를 포함한다.
    - AI 결과(`analysis.result`)는 사고 사실·가해 차량·과실을 확정하지 않는다 (`disclaimer` 참고).
servers:
  - url: http://localhost:3001/api/v1
    description: local
...
security:
  - bearerAuth: []
paths:
...
  /incidents/{id}/submissions:
    post:
      tags: [F5 submissions]
      summary: 증거 영상 제보 생성 + signed upload URL (WITNESS, 알림 받은 사고만)
      description: |
        사고당 제보자 1건. 이미 `UPLOADING`/`ANALYSIS_FAILED` 인 제보가 있으면 같은 행으로 URL 을 재발급(200)하고,
        그 외 상태면 409 `ALREADY_EXISTS`. 동시 요청은 `ON CONFLICT DO NOTHING` 으로 하나만 생성된다.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mime, bytes]
              additionalProperties: false
              properties:
                mime: { type: string, enum: [video/mp4] }
                bytes: { type: integer, minimum: 1 }
                durationSec: { type: number, exclusiveMinimum: 0, maximum: 86400 }
                recordedAt: { type: string, format: date-time }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/SubmissionUpload' }
        '200': { description: 기존 UPLOADING/ANALYSIS_FAILED 제보의 URL 재발급 }
        '403': { $ref: '#/components/responses/ForbiddenRole' }
        '404': { description: 알림을 받지 않은 사고 }
        '409': { description: ALREADY_EXISTS / INVALID_STATE }
        '413': { description: FILE_TOO_LARGE }
        '415': { description: UNSUPPORTED_MEDIA_TYPE }
        '422': { description: VIDEO_UNANALYZABLE (선언한 durationSec 이 허용 범위 밖) }
  /submissions/{id}/complete-upload:
    post:
      tags: [F5 submissions]
      summary: 업로드 완료 확정 (WITNESS). 객체 존재·MIME·MP4 헤더(길이/해상도)·sha256 검증 → UPLOADED
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/Submission' }
        '409': { description: 'INVALID_STATE (UPLOADING 이 아님 / 객체가 아직 없음: message 가 UPLOAD_NOT_FOUND 로 시작)' }
        '413': { description: FILE_TOO_LARGE (실제 객체 크기) }
        '415': { description: UNSUPPORTED_MEDIA_TYPE (객체 content-type 이 video/mp4 아님) }
        '422': { description: 'VIDEO_UNANALYZABLE (MP4 헤더에서 길이/해상도를 읽을 수 없거나 길이가 허용 범위 밖)' }
  /submissions/{id}:
    get:
      tags: [F5 submissions]
      summary: 제보 상세 (제보자, 해당 X, OPERATOR). X 에게 제보자는 `maskedId` 로만 보임
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data: { $ref: '#/components/schemas/Submission' }
        '404': { $ref: '#/components/responses/NotFound' }
  /submissions/{id}/video-url:
    get:
      tags: [F5 submissions]
      summary: 원본 영상 signed URL (제보자 항상, X 는 READY+incidentDetected 이후, OPERATOR)
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          url: { type: string }
                          expiresAt: { type: string, format: date-time }
                          mime: { type: string }
                          sha256: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: INVALID_STATE }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT, description: Supabase access token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
  responses:
...
    ForbiddenRole:
      description: FORBIDDEN_ROLE
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    NotFound:
      description: NOT_FOUND (접근 권한 없는 리소스도 404)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    Envelope:
      type: object
      properties:
        data: {}
        meta:
          type: object
          properties:
            requestId: { type: string }
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            retryable: { type: boolean }
            details: {}
        meta:
          type: object
          properties:
            requestId: { type: string }
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR      # 400
        - UNAUTHENTICATED       # 401
        - FORBIDDEN_ROLE        # 403
        - NOT_FOUND             # 404
        - INVALID_STATE         # 409
        - ALREADY_EXISTS        # 409
        - FILE_TOO_LARGE        # 413
        - UNSUPPORTED_MEDIA_TYPE # 415
        - VIDEO_UNANALYZABLE    # 422 (too short/long, unreadable MP4 header)
        - RATE_LIMITED          # 429
        - INTERNAL_ERROR        # 500
        - PROVIDER_UNAVAILABLE  # 503
...
    SubmissionStatus:
      type: string
      enum: [UPLOADING, UPLOADED, ANALYZING, READY, ANALYSIS_FAILED, SUBMITTED, ADOPTED, REJECTED]
    SubmissionUpload:
      type: object
      properties:
        submissionId: { type: string, format: uuid }
        incidentId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/SubmissionStatus' }
        objectPath: { type: string, example: 'incidents/{incidentId}/submissions/{submissionId}/original.mp4' }
        uploadUrl: { type: string }
        token: { type: string }
        expiresAt: { type: string, format: date-time }
        limits:
          type: object
          properties:
            maxBytes: { type: integer }
            minSeconds: { type: number }
            maxSeconds: { type: number }
            mime: { type: string, const: video/mp4 }
    Submission:
      type: object
      properties:
        id: { type: string, format: uuid }
        incidentId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/SubmissionStatus' }
        mime: { type: string }
        bytes: { type: integer }
        durationSec: { type: [number, 'null'] }
        width: { type: [integer, 'null'] }
        height: { type: [integer, 'null'] }
        sha256: { type: [string, 'null'] }
        recordedAt: { type: [string, 'null'], format: date-time }
        uploadCompletedAt: { type: [string, 'null'], format: date-time }
        objectPath: { type: string, description: 제보자 본인/OPERATOR 에게만 }
        witness:
          type: object
          properties:
            self: { type: boolean }
            maskedId: { type: string, example: '제보자 #1' }
        redactionApplied: { type: boolean, const: false }
        humanReviewed: { type: boolean, const: false }
        analysis:
          oneOf: [{ $ref: '#/components/schemas/Analysis' }, { type: 'null' }]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AnalysisStatus: { type: string, enum: [QUEUED, ANALYZING, FINALIZING, READY, FAILED] }
    AnalysisSource: { type: [string, 'null'], enum: [LIVE, PRERECORDED, null] }
    AnalysisResult:
      type: object
      description: 서버 정규화 결과. `incidentDetected=false` 이면 `incidentTimestampSeconds` 는 null.
      properties:
        incidentDetected: { type: boolean }
        incidentTimestampSeconds: { type: [number, 'null'] }
        incidentTimestampLabel: { type: [string, 'null'], example: '00:12' }
        victimVehicle: { type: [string, 'null'] }
        otherVehicle: { type: [string, 'null'] }
        event: { type: [string, 'null'] }
        relevance: { type: string, enum: [HIGH, MEDIUM, LOW] }
        evidence: { type: array, items: { type: string } }
        videoDurationSec: { type: [number, 'null'] }
        disclaimer: { type: string }
    Analysis:
      type: object
      properties:
        id: { type: string, format: uuid }
        submissionId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/AnalysisStatus' }
        source: { $ref: '#/components/schemas/AnalysisSource' }
        result:
          oneOf: [{ $ref: '#/components/schemas/AnalysisResult' }, { type: 'null' }]
        error:
          oneOf:
            - type: object
              properties:
                code: { type: string, enum: [TIMEOUT, PROVIDER_UNAVAILABLE, PROVIDER_REJECTED, INVALID_RESPONSE] }
                message: { type: [string, 'null'] }
            - type: 'null'
        attempts: { type: integer }
        model: { type: [string, 'null'] }
        promptVersion: { type: [string, 'null'] }
        provider: { type: [string, 'null'] }
        startedAt: { type: [string, 'null'], format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...